UI update.
After table view is reloaded, the table now scrolls to the last sent message.

Bugs: when keyboard is shown, the upper part of message table is moved up, so you can't see top messages.
There are plenty of questions (as of June 2020) for that problem. One solution (taken from https://www.udemy.com/course/ios-13-app-development-bootcamp/learn/lecture/16813360#questions/9497114) is the following:

Step 1: In the loadMessages() function, add the following two lines of code to the block where the cells are being reloaded and the code scrolls to the bottom:
"
func loadMessages() {
...
    DispatchQueue.main.async {
        // add the following line...
        self.tableView.contentInset = UIEdgeInsets.zero
        self.tableView.reloadData()
        // add the following line...
        self.adjustTableViewHeight(for: self.tableView)
        let indexPath = IndexPath(row: self.messages.count - 1, section: 0)
        self.tableView.scrollToRow(at: indexPath, at: .top, animated: false)
    }
...
}
Step 2: Add the following function to adjust the Table View top edge inset if there aren't enough cells to fill the view:

func adjustTableViewHeight(for tableView: UITableView) {
    var totalMessagesHeight = CGFloat(0.0)
    for cell in tableView.visibleCells {
        totalMessagesHeight += cell.bounds.height
    }
    if (totalMessagesHeight < tableView.frame.height) {
        tableView.contentInset = UIEdgeInsets(top: tableView.frame.height - totalMessagesHeight, left: 0, bottom: 0, right: 0)
    }
"